feature: Allow non-calcom domains to run team slug pages
