+++ apps/web/lib/team/[slug]/getServerSideProps.tsx
@@ -1,13 +1,17 @@
 import type { GetServerSidePropsContext } from "next";
 
 import { orgDomainConfig } from "@calcom/features/ee/organizations/lib/orgDomains";
+import {
+  getOrganizationSettings,
+  getVerifiedDomain,
+} from "@calcom/features/ee/organizations/lib/orgSettings";
 import { getFeatureFlag } from "@calcom/features/flags/server/utils";
+import { IS_CALCOM } from "@calcom/lib/constants";
 import { getUserAvatarUrl } from "@calcom/lib/getAvatarUrl";
 import { getBookerBaseUrlSync } from "@calcom/lib/getBookerUrl/client";
 import logger from "@calcom/lib/logger";
 import { markdownToSafeHTML } from "@calcom/lib/markdownToSafeHTML";
 import { getTeamWithMembers } from "@calcom/lib/server/queries/teams";
-import { OrganizationRepository } from "@calcom/lib/server/repository/organization";
 import slugify from "@calcom/lib/slugify";
 import { stripMarkdown } from "@calcom/lib/stripMarkdown";
 import prisma from "@calcom/prisma";
@@ -30,8 +34,13 @@ function getOrgProfileRedirectToVerifiedDomain(
   if (!team.isOrganization) {
     return null;
   }
+  // when this is not on a Cal.com page we don't auto redirect -
+  // good for diagnosis purposes.
+  if (!IS_CALCOM) {
+    return null;
+  }
 
-  const verifiedDomain = OrganizationRepository.utils.getVerifiedDomain(settings);
+  const verifiedDomain = getVerifiedDomain(settings);
 
   if (!settings.orgProfileRedirectsToVerifiedDomain || !verifiedDomain) {
     return null;
@@ -145,7 +154,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
     } as const;
   }
 
-  const organizationSettings = OrganizationRepository.utils.getOrganizationSettings(team);
+  const organizationSettings = getOrganizationSettings(team);
   const allowSEOIndexing = organizationSettings?.allowSEOIndexing ?? false;
 
   const redirectToVerifiedDomain = organizationSettings
